feat: refactor SDK architecture for separation of concerns
Major API restructuring with clear separation between payments, storage coordination, and PDP protocol layers

- Split SynapsePayments into PaymentsService (pure payments) and PandoraService (storage coordination)
- Move storage cost calculations from payments to PandoraService where they belong
- Consolidate PDP components: merge multiple services into unified PDPServer class
- Add PDPVerifier for low-level contract interactions
- Remove storage-specific methods from main Synapse class for cleaner interface
- Introduce FilecoinNetworkType for better type safety across the codebase
- Add removeServiceProvider to PandoraService with URL change detection in setup script

Key improvements:
- PaymentsService now handles only token operations and approvals
- PandoraService manages storage costs, allowances, and proof sets
- PDPServer provides unified interface for all Curio HTTP operations
- Better separation matches external service architecture
- Enhanced type safety with dedicated network type
- Improved deployment tooling with automatic SP re-registration on URL changes

Migration guide:
- Replace synapse.payments.calculateStorageCost() with PandoraService methods
- Use PDPServer instead of separate Upload/Download/Tool services
- PDPService renamed to PandoraService with expanded functionality
- Import individual components from their dedicated paths
