+++ src/index.ts
@@ -4,7 +4,9 @@
 
 export * from './types.js'
 export { Synapse } from './synapse.js'
-export { SynapsePayments } from './payments/index.js'
 export { MockStorageService } from './storage-service.js'
 export { MockUploadTask } from './upload-task.js'
 export * from './utils/index.js'
+export * from './payments/index.js'
+export * from './pandora/index.js'
+export * from './pdp/index.js'
