+++ src/pandora/index.ts
@@ -1,0 +1,9 @@
+// Export Pandora components
+export { PandoraService } from './service.js'
+export type {
+  AddRootsInfo,
+  ApprovedProviderInfo,
+  ComprehensiveProofSetStatus,
+  PendingProviderInfo,
+  ProofSetCreationVerification
+} from './service.js'
