+++ src/pdp/auth.ts
@@ -547,50 +547,3 @@ export class PDPAuthHelper {
     return await this.signer.getAddress()
   }
 }
-
-/**
- * Create authentication blob for PDP service verification
- *
- * Converts an authentication signature into the format expected by PDP services.
- * This is typically the raw signature that will be verified by the smart contract.
- *
- * @param authSignature - The authentication signature from PDPAuthHelper methods
- * @returns Hex-encoded signature suitable for PDP service authentication
- *
- * @example
- * ```typescript
- * const auth = new PDPAuthHelper(contractAddress, signer, chainId)
- * const signature = await auth.signCreateProofSet(0, providerAddress, false)
- * const authBlob = createAuthBlob(signature)
- * // Use authBlob in PDP service API calls
- * ```
- */
-export function createAuthBlob (authSignature: AuthSignature): string {
-  // Return the signature as hex for PDP service verification
-  return authSignature.signature
-}
-
-/**
- * Verify that a signature matches the expected signer
- *
- * Utility function for testing and debugging authentication signatures.
- * Note: This won't work correctly for EIP-712 signatures without the full typed data.
- *
- * @param authSignature - The authentication signature to verify
- * @param expectedSigner - The expected signer's Ethereum address
- * @returns True if the signature was created by the expected signer
- */
-export function verifySignature (
-  authSignature: AuthSignature,
-  expectedSigner: string
-): boolean {
-  // For EIP-712, we can't verify without the full typed data
-  // This is just for compatibility - real verification happens in the contract
-  try {
-    const sig = ethers.Signature.from(authSignature.signature)
-    // Basic check that signature is well-formed
-    return sig.r.length === 66 && sig.s.length === 66 && (sig.v === 27 || sig.v === 28)
-  } catch {
-    return false
-  }
-}
